V2 of the jupyter notebooks
We have a yearly forecast that also multiplies out CVSS vectors for mini-forecasts of those. It has a bug fix in the visualisation that means the forecasts extends a dotted line from the last CY.  The quarterly forecast code has a similar bug fix for visualisation, and some new seasonal orders that can be used with good judgement. The datafetching code has been updated as well. This should allow some of the forecasting team to produce forecasts independently, and expand on the work already done.